Fixes #274 amount must be in minor units

+++ app/code/community/Adyen/Payment/Model/Observer.php
@@ -201,7 +201,7 @@ class Adyen_Payment_Model_Observer {
         }
 
         $adyFields = array(
-            "paymentAmount"     => Mage::helper('adyen')->formatAmount($this->_getCurrentPaymentAmount(), $this->_getCurrentCurrencyCode()),
+            "paymentAmount"     => (int) Mage::helper('adyen')->formatAmount($this->_getCurrentPaymentAmount(), $this->_getCurrentCurrencyCode()),
             "currencyCode"      => $this->_getCurrentCurrencyCode(),
             "merchantReference" => "Get Payment methods",
             "skinCode"          => $skinCode,
